+++ docs-web/.vitepress/theme/index.ts
@@ -56,6 +56,7 @@ import EvTextarea from '../../../packages/evoke-ui/src/components/textarea/index
 import EvSelect from '../../../packages/evoke-ui/src/components/select/index.vue'
 import EvField from '../../../packages/evoke-ui/src/components/field/index.vue'
 import EvSlider from '../../../packages/evoke-ui/src/components/slider/index.vue'
+import EvScrollScene from '../../../packages/evoke-ui/src/components/scroll-scene/index.vue'
 import EvPullRefresh from '../../../packages/evoke-ui/src/components/pull-refresh/index.vue'
 import EvLoadMore from '../../../packages/evoke-ui/src/components/load-more/index.vue'
 import EvActionSheet from '../../../packages/evoke-ui/src/components/action-sheet/index.vue'
@@ -87,7 +88,7 @@ export default {
       EvTimeline, EvComparisonTable, EvCta, EvNewsletter, EvLogoCloud,
       EvVideo, EvAudio, EvContactForm, EvCarousel, EvArticleCard, EvProfileCard,
       EvBorderBeam, EvExecCard, EvArticle, EvImageWall, EvImagePreview, EvWaterfall, EvModal,
-      EvMarquee, EvInput, EvTextarea, EvSelect, EvField, EvSlider,
+      EvMarquee, EvInput, EvTextarea, EvSelect, EvField, EvSlider, EvScrollScene,
       EvPullRefresh, EvLoadMore, EvActionSheet, EvTabbar, EvTabbarItem, EvNavBar,
       DemoBlock, MobileStage,
     }